LayoutTests/fast/profiler/multiple-frames-expected.txt

CONSOLE MESSAGE: line 59: Other window executing JavaScript
CONSOLE MESSAGE: line 75: Thread_1  0
CONSOLE MESSAGE: line 75:  (program) multiple-frames.html 8
CONSOLE MESSAGE: line 75:   (program) multiple-frames.html 4
CONSOLE MESSAGE: line 75:   (program) other-frame.html 3
CONSOLE MESSAGE: line 75:   onload multiple-frames.html 24
CONSOLE MESSAGE: line 75:    testStart multiple-frames.html 10
CONSOLE MESSAGE: line 75:     getElementById  0
CONSOLE MESSAGE: line 75:     functionInOtherFrame other-frame.html 4
CONSOLE MESSAGE: line 75:      functionInParentFrame multiple-frames.html 18
CONSOLE MESSAGE: line 75:     insertGivenText profiler-test-JS-resources.js 9
CONSOLE MESSAGE: line 75:      createElement  0
CONSOLE MESSAGE: line 75:      createTextNode  0
CONSOLE MESSAGE: line 75:      appendChild  0
CONSOLE MESSAGE: line 75:      getElementById  0
CONSOLE MESSAGE: line 75:     endTest profiler-test-JS-resources.js 1
CONSOLE MESSAGE: line 75:  (idle)  0
This test has JS executing in another window. 

To run this test manually, load it in the browser then load the WebInspector and look at the profile. In the profile there should be a call to functionInOtherFrame().
Pi
